Handle show hide ourselves
Remove the data-toggle="dropdown" and no longer rely on bootstrap's
JavaScript. Instead we should to that our selves. This is because the
bootstrap plugin doesn't off enough hooks to prevent bubbling and would
make the dropdown close when we didn't want them to. We now manage it's
state with ember.