Add single select all/none feature
The split All/None button group was meant for feature parity with the
bootstrap-select project this was inspired from. However, others have
noted that it looks a bit ugly and a single toggle button would be
nicer.

This adds a single button that toggles the selection contextually. The
old functionality can be enabled setting splitAllNoneButtons to true.