Refactor the summaryMessage logic
The logic was tripping over it's own feet because the I18n setup and how
it handles the translatable properties. The only way out of the rats
nest was to use summaryMessageKey instead and do our own translation in
a computed property. Documentation will have to be clear on this
discrepancy.