Add promise resolution for bound properties
Fixes issue #29

Ember.Data relations are actually promises and a component needs to
manage the future value appropriately. Looking into how Ember's Select
View handles this it defers returning a value in the getter till the
promise is complete and then using it's own setter to set the value. If
the value is not a promise it is returned as normal (immediately).