Set keyboard shortcuts scoped
Fixes issue #45
Closes pull request #46

The mousetrap library used by ember-keyboard-shortcuts (mousetrap)
defaults to capturing the keyup event on document. This patch forces the
events to be scoped to the component that they are defined on.

When it was attached to the document it would trap the Up, Down, Tab,
Shift-Tab, and Enter keys for any ember view that used the
keyboard-select-picker. Not helpful. In the future we may wish to move
to rolling our own jQuery keyboard logic to remove the dependency on
ember-keyboard-shortcuts.