Add an itemId
When iterating over the nestedGroupContentList there wasn't an easy way
to tell which index of the contentList an iteration was on. Now we can
use the itemId to know how to look up an specific item by index. It
helps in the reverse look up to know which DOM element is associated
with an particular item using the `data-itemid` tag.