Assign dropup class automatically
One problem with the menu is that if it is near the bottom of the page
the CSS has it larger then the bottom of the page cutting of a portion
of the menu. It is difficult to select something when it is clipped off
and you can not scroll to it.

We calculate the relative sizing every time the page resizes or scrolls
so that we can attach the 'dropup' class. This class tells bootstrap to
make the menu pop up instead of down allowing selects that are near the
bottom to be visible.