Add option to override the summaryMessage
This was sort of working. It would allow you to customize the
summary message only when there were multiple values selected.

Now if you add the summaryMessage property it will just use that
regardless. It is up to the user to make it a computed property if you
want it dynamic.